requirements.txt
# requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
joblib==1.3.0
plotly==5.17.0
fpdf2==2.7.4
python-dotenv==1.0.0
SQLAlchemy==2.0.23
WTForms==3.1.1
Werkzeug==3.0.1